v1/v2 setup deal benches
