+++ types/src/dkg/real_dkg/rounding/mod.rs
@@ -371,3 +371,135 @@ pub static DEFAULT_RECONSTRUCT_THRESHOLD: Lazy<U64F64> =
 
 pub static DEFAULT_FAST_PATH_SECRECY_THRESHOLD: Lazy<U64F64> =
     Lazy::new(|| U64F64::from_num(2) / U64F64::from_num(3));
+
+pub const MAINNET_STAKES: [u64; 129] = [
+    145363920367444000,
+    100779935896493000,
+    134154255721034000,
+    134234783226671000,
+    103686549105772000,
+    23681356495150300,
+    134577645972875000,
+    134580712197205000,
+    10857449995312600,
+    105977831715137000,
+    120872333189108000,
+    102057954967375000,
+    25968006480319300,
+    150210261808047000,
+    11047428320304400,
+    134184685206018000,
+    128117795425337000,
+    12497398680912800,
+    50533200268704000,
+    10856898438192700,
+    134176595090811000,
+    60011869592362800,
+    39694335719301200,
+    12863458468719700,
+    11046541966772300,
+    94427102742955200,
+    58714132394437700,
+    40145680094791300,
+    100137028609146000,
+    111809600151787000,
+    114998912365121000,
+    17951622559336400,
+    10857216258421800,
+    94429160256130900,
+    11047450111666100,
+    10856891072965200,
+    10857020952663600,
+    162161975531451000,
+    104073248041097000,
+    10857355385008300,
+    10901683472171900,
+    50198365896562800,
+    134182311143049000,
+    10857000567453500,
+    134562155405120000,
+    11046814332439800,
+    60322390260321900,
+    101055534219498000,
+    114872371828424000,
+    10929001624435000,
+    106067388838015000,
+    100152796096971000,
+    54964069016926100,
+    34040647636753800,
+    102049697282067000,
+    10856874306102900,
+    134514110645862000,
+    70499650588096400,
+    134224065841861000,
+    10857370648417500,
+    28409091651485400,
+    64302033587393400,
+    16659350234613100,
+    112568696851409000,
+    10857782940276500,
+    200882335168720000,
+    10856846459964200,
+    10856305691600600,
+    10856576121655500,
+    123961368695808000,
+    20275491671732600,
+    112069796227716000,
+    148078356637657000,
+    76893226659146600,
+    135298123702389000,
+    10856788596777500,
+    107821720522194000,
+    134626203055928000,
+    106466193065101000,
+    102103040930732000,
+    62682920098289700,
+    26223235705449200,
+    134234424849999000,
+    150210282994581000,
+    134913703983987000,
+    10857227273097400,
+    57413947132891200,
+    10900450777364100,
+    12022049676664200,
+    11047053887431400,
+    10857590490261700,
+    10857257627847700,
+    26223774458854400,
+    145363467800313000,
+    49332020110088600,
+    11047476999099800,
+    126201751573407000,
+    150458532010203000,
+    10856470759531000,
+    10857203409232300,
+    11047327948099800,
+    10856521489540500,
+    99511242999587700,
+    74202213386306600,
+    11047051193450900,
+    32601393370365500,
+    70855459554627300,
+    10857401127909800,
+    25271130862928600,
+    18684565615586300,
+    10876016328079300,
+    95866473180260000,
+    10857461985291600,
+    10857176446687100,
+    17291414467856800,
+    47528452645556600,
+    17051109168257700,
+    134912746458206000,
+    150461059796590000,
+    116315225160302000,
+    10855453497812600,
+    100865713263752000,
+    10928177475521100,
+    124293961561247000,
+    26223786196810300,
+    39221522777191400,
+    73810826128117800,
+    53685896908423700,
+    40216803848486900,
+];
